Added code

+++ MRF/MRF.Models/Models/Candidatedetails.cs
@@ -12,7 +12,7 @@ public class Candidatedetails
 
     public string EmailId { get; set; } = null!;
 
-    public int ContactNo { get; set; } 
+    public  long ContactNo { get; set; } 
 
     public string ResumePath { get; set; } = null!;
